requirements.txt
discord.py-self
groq
python-dotenv